support for tty login managers
TTY login managers that switch to graphical login later are handled by tracking for
session "Type" property change to a graphical one. This approach may not work
for explicit startx/startkde/... invocations for a session start in tty login because
those may not change the "Type" property of the session but should work for compositors
that update the session type.