Feature: support for tty login managers (#3)
TTY login managers that switch to graphical login later are handled by tracking for
session "Type" property change to a graphical one. This approach may not work
for explicit session start in tty login for compositors like KDE plasma because
those may not change the "Type" property of the session but works for compositors
that update the systemd-logind session type (like sway, hyprland etc).